requirements.txt
sentence-transformers 
faiss-cpu 
tiktoken 
pandas 
requests 
python-dotenv
transformers 
torch
mediacloud	
pydantic-settings